Move struct ds_worker to ds_worker.h
